+++ app/browser/tools.js
@@ -1,0 +1,14 @@
+function validateFile() {
+  window = BrowserWindow.getFocusedWindow();
+  window.webContents.send('validate');
+}
+
+function generateSchemaFromHeader() {
+  window = BrowserWindow.getFocusedWindow();
+  window.webContents.send('schemaFromHeaders');
+}
+
+function fixRaggedRowsFile() {
+  window = BrowserWindow.getFocusedWindow();
+  window.webContents.send('ragged_rows');
+}
